+++ pallets/staking/src/tests.rs
@@ -14,8 +14,7 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 use crate::{
-    mock::*, tests::RuntimeEvent, Error, IsValidatorSynced, NextSignerInfo, NextSigners,
-    ServerInfo, Signers, ThresholdToStash,
+    mock::*, tests::RuntimeEvent, Error, NextSignerInfo, NextSigners, ServerInfo, Signers,
 };
 use codec::Encode;
 use frame_support::{assert_noop, assert_ok};
@@ -49,8 +48,6 @@ fn basic_setup_works() {
         );
         assert_eq!(Staking::threshold_to_stash(7).unwrap(), 5);
         assert_eq!(Staking::threshold_to_stash(8).unwrap(), 6);
-        assert!(Staking::is_validator_synced(5));
-        assert!(Staking::is_validator_synced(6));
     });
 }
 
@@ -84,7 +81,7 @@ fn it_takes_in_an_endpoint() {
         let server_info = ServerInfo {
             tss_account: 3,
             x25519_public_key: NULL_ARR,
-            endpoint: vec![20, 20, 20, 20],
+            endpoint: vec![20; 26],
             provisioning_certification_key: BoundedVec::with_max_capacity(),
         };
         assert_noop!(
@@ -325,8 +322,6 @@ fn it_deletes_when_no_bond_left() {
             VALID_QUOTE.to_vec(),
         ));
 
-        IsValidatorSynced::<Test>::insert(2, true);
-
         let ServerInfo { tss_account, endpoint, .. } = Staking::threshold_server(2).unwrap();
         assert_eq!(endpoint, vec![20]);
         assert_eq!(tss_account, 3);
@@ -359,8 +354,6 @@ fn it_deletes_when_no_bond_left() {
         lock = Balances::locks(2);
         assert_eq!(lock[0].amount, 50);
         assert_eq!(lock.len(), 1);
-        // validator still synced
-        assert_eq!(Staking::is_validator_synced(2), true);
 
         let ServerInfo { tss_account, endpoint, .. } = Staking::threshold_server(2).unwrap();
         assert_eq!(endpoint, vec![20]);
@@ -377,8 +370,6 @@ fn it_deletes_when_no_bond_left() {
         assert_eq!(lock.len(), 0);
         assert_eq!(Staking::threshold_server(2), None);
         assert_eq!(Staking::threshold_to_stash(3), None);
-        // validator no longer synced
-        assert_eq!(Staking::is_validator_synced(2), false);
 
         assert_ok!(FrameStaking::bond(
             RuntimeOrigin::signed(7),
@@ -426,21 +417,6 @@ fn it_deletes_when_no_bond_left() {
 }
 
 #[test]
-fn it_declares_synced() {
-    new_test_ext().execute_with(|| {
-        assert_noop!(
-            Staking::declare_synced(RuntimeOrigin::signed(5), true),
-            Error::<Test>::NoThresholdKey
-        );
-
-        ThresholdToStash::<Test>::insert(5, 5);
-
-        assert_ok!(Staking::declare_synced(RuntimeOrigin::signed(5), true));
-        assert!(Staking::is_validator_synced(5));
-    });
-}
-
-#[test]
 fn it_tests_new_session_handler() {
     new_test_ext().execute_with(|| {
         // Start with current validators as 5 and 6 based off the Mock `GenesisConfig`.
@@ -462,19 +438,18 @@ fn it_tests_new_session_handler() {
             last_session_change: 0,
         });
 
-        assert_ok!(Staking::new_session_handler(&[1, 2, 3]));
-        // takes signers original (5,6) pops off first 5, adds (fake randomness in mock so adds 1)
+        assert_ok!(Staking::new_session_handler(&[1, 5, 6]));
+        // takes signers original (5,6) pops off one and adds in new validator
         assert_eq!(Staking::next_signers().unwrap().next_signers, vec![6, 1]);
-
         assert_eq!(
             Staking::reshare_data().block_number,
             101,
             "Check reshare block start at 100 + 1"
         );
         assert_eq!(
-            Staking::reshare_data().new_signer,
-            1u64.encode(),
-            "Check reshare next signer up is 1"
+            Staking::reshare_data().new_signers,
+            vec![1u64.encode()],
+            "Check reshare next signer up is 3"
         );
         assert_eq!(
             Staking::jump_start_progress().parent_key_threshold,
@@ -487,11 +462,6 @@ fn it_tests_new_session_handler() {
             101,
             "Check reshare block start at 100 + 1"
         );
-        assert_eq!(
-            Staking::reshare_data().new_signer,
-            1u64.encode(),
-            "Check reshare next signer up is 1"
-        );
 
         assert_ok!(Staking::new_session_handler(&[6, 5, 3]));
         // takes 3 and leaves 5 and 6 since already in signer group
@@ -500,6 +470,42 @@ fn it_tests_new_session_handler() {
         assert_ok!(Staking::new_session_handler(&[1]));
         // does nothing as not enough validators
         assert_eq!(Staking::next_signers().unwrap().next_signers, vec![6, 3]);
+
+        // reduce threshold to make sure next signers does not drop > then threshold of current signers
+        pallet_parameters::SignersInfo::<Test>::put(SignersSize {
+            total_signers: 2,
+            threshold: 1,
+            last_session_change: 0,
+        });
+
+        assert_ok!(Staking::new_session_handler(&[1, 2, 3]));
+        assert_eq!(Staking::next_signers().unwrap().next_signers, vec![5, 1]);
+    });
+}
+
+#[test]
+fn it_tests_new_session_handler_truncating() {
+    new_test_ext().execute_with(|| {
+        // Start with current validators as 7 and 8 based off the Mock `GenesisConfig`.
+        Signers::<Test>::put(vec![7, 8]);
+        System::set_block_number(100);
+        pallet_parameters::SignersInfo::<Test>::put(SignersSize {
+            total_signers: 2,
+            threshold: 2,
+            last_session_change: 0,
+        });
+        // test truncates none if t and n = 0
+        assert_ok!(Staking::new_session_handler(&[1, 2, 3]));
+        assert_eq!(Staking::next_signers().unwrap().next_signers, vec![7, 8]);
+
+        pallet_parameters::SignersInfo::<Test>::put(SignersSize {
+            total_signers: 2,
+            threshold: 1,
+            last_session_change: 0,
+        });
+        // test truncates 1 if n - t = 1
+        assert_ok!(Staking::new_session_handler(&[1, 2, 3]));
+        assert_eq!(Staking::next_signers().unwrap().next_signers, vec![7, 1]);
     });
 }